fixing header bugs
- changed Header component to server side
- fixed flash bug inside Header component
- changed HeaderIcon to server side
- created new independent component for Burger to use client side rendering
- refactored and deleted unnecessary code
- replaced client side session check to server side getSession inside LogOutButton
