+++ components/headerFooterContent/Header.tsx
@@ -1,9 +1,6 @@
-"use client";
 
-import { GiHamburgerMenu } from "react-icons/gi";
+
 import { Navigation } from "./Navigation";
-import { MobileMenu } from "../headerFooterContent/MobileMenu";
-import { useState } from "react";
 import { HeaderIcon } from "./HeaderIcon";
 import Link from "next/link";
 import { LogOutButton } from "../logOutButton/LogOutButton";
@@ -11,17 +8,20 @@ import ThemeMode from "./ThemeMode";
 import { LangButton } from "../languageChange/LangButton";
 import Image from "next/image";
 import LogoImage from "../../public/assets/image/LogoImage.png";
+import { Claims } from "@auth0/nextjs-auth0";
+import Burger from "../Burger";
 
-export default function Header() {
-  const [open, setOpen] = useState(false);
+export default function Header({ session }: { session: Claims | undefined }) {
+  // const [open, setOpen] = useState(false);
 
-  const handleCheck = () => {
-    setOpen(true);
-  };
+  // const handleCheck = () => {
+  //   setOpen(true);
+  // };
+
   return (
     <header className="w-full bg-[#dbe0da] py-4 px-0 border-b border-[#111827] dark:bg-[#264653] dark:border-[#ffffff] fixed top-0 left-0 z-10 ">
       <div className="w-full  p-[15px] flex items-center justify-center mobile">
-        <LogOutButton />
+        <LogOutButton session={session} />
         <LangButton />
         <ThemeMode />
       </div>
@@ -35,16 +35,12 @@ export default function Header() {
             height={5}
           />
         </Link>
-        <GiHamburgerMenu
-          className="w-7 h-7 text-[#264653] cursor-pointer lg:hidden"
-          onClick={handleCheck}
-        />
         <div className="hidden lg:block">
           <Navigation marginLeft="30px" display="flex" />
         </div>
         <HeaderIcon marginRight="20px" alignItems="center" />
+        <Burger />
       </div>
-      {open && <MobileMenu setOpen={setOpen} />}
     </header>
   );
 }
