+++ components/headerFooterContent/MobileMenu.tsx
@@ -2,7 +2,7 @@ import { IoClose } from "react-icons/io5";
 import { SocialLinks } from "./SocialLinks";
 import { Navigation } from "./Navigation";
 
-const MobileMenu = ({ setOpen }) => {
+const MobileMenu = ({ setOpen }: { setOpen: React.Dispatch<React.SetStateAction<boolean>> }) => {
   return (
     <div className="fixed top-0 right-0 w-1/2 h-screen bg-gray-800 p-8 flex items-start justify-start flex-col z-10">
       <div
@@ -14,7 +14,7 @@ const MobileMenu = ({ setOpen }) => {
       <div>
         <Navigation
           flexDirection="column"
-          alingItems="flex-start"
+          alignItems="flex-start"
           marginBottom="20px"
           color="#FFFFFF"
           fontSize="25px"
