+++ components/logOutButton/LogOutButton.tsx
@@ -2,14 +2,13 @@
 
 // import { useRouter } from "next/navigation";
 // import { handleLogout } from "@/helper/userLogOut";
-import { useUser } from '@auth0/nextjs-auth0/client';
 import { HiOutlineLogout } from "react-icons/hi";
 import { IoLogIn } from "react-icons/io5";
-import Link from 'next/link';
+import { Claims } from '@auth0/nextjs-auth0';
 
-const LogOutButton = () => {
+const LogOutButton = ({ session }: { session: Claims | undefined }) => {
   // const router = useRouter();
-  const { user } = useUser();
+  const user = session;
 
   // const handleClick = async () => {
   //   await handleLogout();
@@ -20,15 +19,15 @@ const LogOutButton = () => {
   return (
     <div>
       {user ? (
-        <Link href="/api/auth/logout"><HiOutlineLogout
+        <a href="/api/auth/logout"><HiOutlineLogout
           className="w-[20px] h-[20px] text-[#264653] cursor-pointer dark:text-[#ffffff] "
-        /></Link>
+        /></a>
       ) : (
-        <Link href="/api/auth/login">
+        <a href="/api/auth/login">
           <IoLogIn
             className="w-[20px] h-[20px] text-[#264653] cursor-pointer dark:text-[#ffffff] "
           />
-        </Link>
+        </a>
       )}
     </div>
   );
